requirements.txt
matplotlib
tensorflow_gpu
